requirements.txt
python-telegram-bot==21.6
google-generativeai==0.7.2
python-docx==1.1.2
